+++ packages/ragbits-document-search/CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 
 ## Unreleased
-
+- reranker based on relevance to given query (#443)
 - move sources from ragbits-document-search to ragbits-core (#496)
 - fix union types validation in element enricher (#499)